Added relations between collections

+++ documents/question/question.go
@@ -8,14 +8,15 @@ import (
 
 type Question struct {
 	ID           primitive.ObjectID `bson:"_id"`
-	CreatedAt    time.Time          `bson:"createdAt"`
-	UpdatedAt    time.Time          `bson:"updatedAt"`
-	Question     string             `bson:"question"`
-	Options      []string           `bson:"options"`
-	Answer       string             `bson:"answer"`
-	QuestionType string             `bson:"questionType"`
-	Marks        int                `bson:"marks"`
-	Author       string             `bson:"author"`
-	Deleted      bool               `bson:"deleted"`
-	DeletedBy    string             `bson:"deletedBy"`
+	// ID           int64     `bson:"_id"`
+	CreatedAt    time.Time `bson:"createdAt"`
+	UpdatedAt    time.Time `bson:"updatedAt"`
+	Question     string    `bson:"question"`
+	Options      []string  `bson:"options"`
+	Answer       string    `bson:"answer"`
+	QuestionType string    `bson:"questionType"`
+	Marks        float32       `bson:"marks"`
+	Author       string    `bson:"author"`
+	Deleted      bool      `bson:"deleted"`
+	DeletedBy    string    `bson:"deletedBy"`
 }
